+++ Biblioteca/Controllers/ClienteController.cs
@@ -3,6 +3,7 @@ using Biblioteca.ViewModels;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Web;
 using System.Web.Mvc;
 
@@ -24,8 +25,15 @@ namespace Biblioteca.Controllers
         public ActionResult Index()
         {
             var viewModel = _context.Users.ToList();
-
-            return View(viewModel);
+            if (User.IsInRole("CanManageCustomers"))
+            {
+                return View(viewModel);
+            }
+            else
+            {
+                return View("IndexReadOnly", viewModel);
+            }
+            
         }
 
         public ActionResult Details(int id)
@@ -35,6 +43,10 @@ namespace Biblioteca.Controllers
             {
                 return HttpNotFound();
             }
+            else if (User.IsInRole("CanManageCustomers") == false)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+            }
             else
             {
                 return View(user);
@@ -43,6 +55,9 @@ namespace Biblioteca.Controllers
 
         public ActionResult New()
         {
+            if(User.IsInRole("CanManageCustomers") == false){
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+            }
             var user = new Cliente();
             return View("Edit", user);
         }
@@ -52,6 +67,10 @@ namespace Biblioteca.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Save(Cliente user)
         {
+            if(User.IsInRole("CanManageCustomers") == false){
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+            }
+
             if (!ModelState.IsValid)
             {
                 return View("Edit");
@@ -80,6 +99,10 @@ namespace Biblioteca.Controllers
 
         public ActionResult Edit(int id)
         {
+            if(User.IsInRole("CanManageCustomers") == false){
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+            }
+
             var user = _context.Users.SingleOrDefault(c => c.Id == id);
 
             if (user == null)
@@ -92,6 +115,10 @@ namespace Biblioteca.Controllers
 
         public ActionResult Remove(int id)
         {
+            if(User.IsInRole("CanManageCustomers") == false){
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+            }
+
             var user = _context.Users.Single(m => m.Id == id);
 
             if (user != null) _context.Users.Remove(user);
